pls
Former-commit-id: 33e701335d274dff6bcdb320a9006f103e4b934b

+++ src/gameengine/actors/properties/AddMoneyCycleProperty.java
@@ -22,7 +22,7 @@ public class AddMoneyCycleProperty<G extends ReadAndDamageGrid> extends CyclePro
 	
 	@Override
 	public void action(G grid, Integer actorID) {
-		grid.getWriteableGameStatus().addMoney(myMoney);
+		grid.getWriteableGameStatus().addMoney(myMoney.intValue());
 	}
 
 }
